requirements_local.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.10
django-cors-headers==3.10.1
djangorestframework==3.13.1
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
typing_extensions==4.0.1
